refactor(ir): move the AMD architecture table where both backends reach it
`AMDArchitecture` and `AmdWmma` describe AMD hardware, not C++ codegen,
and the LLVM backend needs the same answers. Neither backend crate can
depend on the other, so the table moves down to the crate they share.

A pure relocation: `cubecl-cpp` re-exports both names, and keeps the
`Architecture` impl, whose `is_wmma_capable`/`is_mfma_capable` are about
what its own rocwmma path can emit rather than about the hardware.
